chore: goal-section 파일 분리
